+++ src/mermaidChartVSCode.ts
